backend/requirements.txt
flask
flask-cors
pymongo
scikit-learn
pandas
numpy
nltk
langdetect
joblib
newspaper3k
requests